requirements.txt
redis>=5.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
transformers>=4.30.0
torch>=2.0.0
Pillow>=9.0.0
python-dotenv>=1.0.0
timm>=0.9.0
